[all] License and copyright updates
Due to some errors within Arm, the materials in asllib were released
with the wrong copyright attribution and license. As part of
rectifying this, the original material was removed from the
repository, and has now been replaced.

Arm would ask that (1) if you have copies with the original copyright
attribution and license that these are destroyed, and (2) if you
distributed the original material to any third parties, that you
provide this message to such parties.

Arm recognizes that a copy of the original material will unavoidably
remain in the git revision control system, and will thus persist in
git clones of the repository.  Such copies will unavoidably maintain
the wrong copyright attribution and license, but they need not be
destroyed.

Arm apologizes for the inconvenience caused by this matter.